// src/api/generated/models/selectionProcessStageEnum.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Brew
 * OpenAPI spec version: 0.0.1
 */

export type SelectionProcessStageEnum = typeof SelectionProcessStageEnum[keyof typeof SelectionProcessStageEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SelectionProcessStageEnum = {
  awaiting_resume_screening: 'awaiting_resume_screening',
  resume_screening: 'resume_screening',
  awaiting_screening_questions: 'awaiting_screening_questions',
  screening_questions: 'screening_questions',
  awaiting_interview: 'awaiting_interview',
  interview: 'interview',
  awaiting_technical_challenge: 'awaiting_technical_challenge',
  technical_challenge: 'technical_challenge',
  awaiting_live_coding: 'awaiting_live_coding',
  live_coding: 'live_coding',
  awaiting_language_assessment: 'awaiting_language_assessment',
  language_assessment: 'language_assessment',
  awaiting_portfolio_review: 'awaiting_portfolio_review',
  portfolio_review: 'portfolio_review',
} as const;